+++ example/lib/text_theme_demo.dart
@@ -11,16 +11,12 @@ class TextThemeDemo extends StatelessWidget {
     if (useSystemChineseFont) {
       return Theme(
         data: ThemeData(
-          textTheme: SystemChineseFont.textTheme
-              .merge(
-                // Custom text theme
-                const TextTheme(
-                  displaySmall: TextStyle(fontStyle: FontStyle.italic),
-                  displayMedium: TextStyle(fontStyle: FontStyle.italic),
-                  displayLarge: TextStyle(fontStyle: FontStyle.italic),
-                ),
-              ) // Apply change to the entire fontTheme
-              .apply(fontFamily: 'LatoLato'),
+          textTheme: const TextTheme(
+            headlineLarge: TextStyle(fontStyle: FontStyle.italic),
+            displaySmall: TextStyle(fontStyle: FontStyle.italic),
+            displayMedium: TextStyle(fontStyle: FontStyle.italic),
+            displayLarge: TextStyle(fontStyle: FontStyle.italic),
+          ).apply(fontFamily: 'LatoLato').useSystemChineseFont(),
         ),
         child: const ThemedTextList(),
       );